Use describeTransformerTests for naming-convention to verify both modes deliver the same results
